pyproject.toml
[project]
name = "bbinternal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "load-dotenv>=0.1.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
